Extract uri matcher
